requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
playwright==1.58.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.2
pyee==13.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.6.3
wakepy==1.0.0
yarl==1.22.0
